PythonService/requirements.txt
numpy==1.16.3
Flask==1.0.3
pdf2image==1.6.0
pymongo==3.9.0
APScheduler==3.6.1
requests==2.22.0
pandas==0.24.2
opencv_python==3.4.5.20
flask_restplus==0.12.1
Pillow==6.1.0
protobuf==3.9.1
Flask_api==1.1
prettytable==0.7.2
matplotlib
scipy
google-cloud-storage
google-cloud-vision